29-apr-2005 Library creation scripts for Linux gcc etc... introduced.             Argument 'float' changed to 'double' for function isnan in amanda.c             to match gcc definition and setting of unused variable 'pi' removed             from IceF2k.cxx.

+++ RALICE/icepack/iceconvert/amanda.c
@@ -37,7 +37,7 @@ long rdmc_nint(double a)
 /*******************************/
 /* Work around missing isnan() */
 /*******************************/
-int isnan(float r)
+int isnan(double r)
 {
  return 0;
 }
